add unit test

+++ inc/Engine/Optimization/RUCSS/Admin/Database.php
@@ -87,12 +87,17 @@ class Database {
 	 *
 	 * @since 3.12
 	 *
-	 * @return void
+	 * @return bool
 	 */
-	public function drop_resources_table() {
+	public function drop_resources_table(): bool {
 		global $wpdb;
 
-		$wpdb->query( "DROP TABLE {$wpdb->prefix}wpr_rucss_resources" ); // phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.DirectDatabaseQuery.SchemaChange
-		delete_option( 'wpr_rucss_resources_version' );
+		$result = $wpdb->query( "DROP TABLE {$wpdb->prefix}wpr_rucss_resources" ); // phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.DirectDatabaseQuery.SchemaChange
+
+		if ( false === $result ) {
+			return false;
+		}
+
+		return delete_option( 'wpr_rucss_resources_version' );
 	}
 }
